+++ lib/config/pagy.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# Pagy initializer file (5.3.1)
+# Pagy initializer file (5.4.0)
 # Customize only what you really need and notice that Pagy works also without any of the following lines.
 # Should you just cherry pick part of this file, please maintain the require-order of the extras
 
@@ -43,14 +43,12 @@
 # Calendar extra: Paginate a collection by calendar Time unit (year, month, week or day)
 # See https://ddnexus.github.io/pagy/extras/calendar
 # require 'pagy/extras/calendar'
-# Pagy::DEFAULT[:local_minmax] = []          # Min and max local Time period must be set by the user (better not not as default)
-# Pagy::DEFAULT[:unit]         = :month      # Time unit allowed %i[year month week day]
-# Pagy::DEFAULT[:week_offset]  = 0           # Day offset from Sunday (0: Sunday; 1: Monday;... 6: Saturday)
-# Pagy::DEFAULT[:order]         = :asc        # Time direction of pagination
-# Pagy::DEFAULT[:year_format]  = '%Y'        # strftime format for :year unit
-# Pagy::DEFAULT[:month_format] = '%Y-%m'     # strftime format for :month unit
-# Pagy::DEFAULT[:week_format]  = '%Y-%W'     # strftime format for :week unit
-# Pagy::DEFAULT[:day_format]   = '%Y-%m-%d'  # strftime format for :day unit
+# DEFAULT[:year_format]  = '%Y'        # strftime format for :year unit
+# DEFAULT[:month_format] = '%Y-%m'     # strftime format for :month unit
+# DEFAULT[:week_format]  = '%Y-%W'     # strftime format for :week unit
+# DEFAULT[:day_format]   = '%Y-%m-%d'  # strftime format for :day unit
+# DEFAULT[:week_offset]  = 0           # Day offset from Sunday (0: Sunday; 1: Monday;... 6: Saturday)
+# DEFAULT[:time_order]   = :asc        # Time direction of pagination
 
 # Countless extra: Paginate without any count, saving one query per rendering
 # See https://ddnexus.github.io/pagy/extras/countless