+++ test/pagy/extras/frontend_helpers_json_test.rb
@@ -27,8 +27,9 @@ describe 'pagy/extras/frontend_helpers_json' do
   describe 'Calendar sequels and label_sequels' do
     it 'generate the labels for the sequels' do
       steps = { 0 => [1, 2, 2, 1], 600 => [1, 3, 3, 1] }
-      pagy = Pagy::Calendar.new(local_minmax: [Time.new(2021, 10, 21, 13, 18, 23, 0), Time.new(2023, 11, 13, 15, 43, 40, 0)],
-                                steps: steps, page: 3)
+      pagy = Pagy::Calendar.create(:month,
+                                   minmax: [Time.new(2021, 10, 21, 13, 18, 23, 0), Time.new(2023, 11, 13, 15, 43, 40, 0)],
+                                   steps: steps, page: 3)
       _(pagy.sequels).must_rematch
       _(pagy.label_sequels).must_rematch
     end